Added close window test
